+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `cautious_delete` to Mixin class, initial implementation. #711
 - Add `deprecation_factory` to facilitate table migration. #717
 - Add Spyglass logger. #730
+- IntervalList: Add secondary key `pipeline` #742
 
 ### Pipelines
 
@@ -23,7 +24,13 @@
 - Position:
     - Refactor input validation in DLC pipeline. #688
     - DLC path handling from config, and normalize naming convention. #722
-- IntervalList: Add secondary key `pipeline` #742
+- Decoding:
+    - Add `decoding` pipeline V1. #731
+    - Add a table to store the decoding results #731
+    - Use the new `non_local_detector` package for decoding #731
+    - Allow multiple spike waveform features for clusterelss decoding #731
+    - Reorder notebooks #731
+
 
 ## [0.4.3] (November 7, 2023)